added online resources specific page container and refactored existing pgs
