+++ src/pages/research/collections/online-resources/[slug].js
@@ -16,13 +16,8 @@ import {
 import { withRedux } from './../../../../redux/withRedux';
 // Components
 import { SkeletonLoader } from '@nypl/design-system-react-components';
-import PageContainer from './../../../../components/shared/layouts/PageContainer';
-import RightRail from './../../../../components/shared/RightRail';
-import SearchHeader from './../../../../components/shared/SearchHeader';
-import SearchForm from './../../../../components/online-resources/SearchForm';
+import PageContainer from './../../../../components/online-resources/layouts/PageContainer';
 import OnlineResourceCard from './../../../../components/online-resources/OnlineResourceCard';
-import SidebarMenus from './../../../../components/online-resources/SidebarMenus';
-import Hero from './../../../../components/online-resources/Hero';
 // Utils
 import { ONLINE_RESOURCES_BASE_PATH } from './../../../../utils/config';
 
@@ -62,20 +57,8 @@ function OnlineResourceSlug() {
   if (loading || !data) {
     return (
       <PageContainer
-        wrapperClass='nypl--research'
-        contentHeader={
-          <Fragment>
-            <Hero />
-            <SearchHeader>
-              <SearchForm />
-            </SearchHeader>
-          </Fragment>
-        }
         contentPrimary={
           <SkeletonLoader />
-        }
-        contentBottom={
-          <RightRail />
         }
       />
     );
@@ -88,15 +71,6 @@ function OnlineResourceSlug() {
         description: `${data.searchDocument.name}`,
         url: `https://www.nypl.org${ONLINE_RESOURCES_BASE_PATH}/${slug}`
       }}
-      wrapperClass='nypl--research'
-      contentHeader={
-        <Fragment>
-          <Hero />
-          <SearchHeader>
-            <SearchForm />
-          </SearchHeader>
-        </Fragment>
-      }
       contentPrimary={
         <Fragment>
           <Link href={`${ONLINE_RESOURCES_BASE_PATH}`}>
@@ -104,14 +78,6 @@ function OnlineResourceSlug() {
           </Link>
           <OnlineResourceCard item={data.searchDocument} />
         </Fragment>
-      }
-      showSidebar={true}
-      sidebarSide='right'
-      contentSecondary={
-        <SidebarMenus />
-      }
-      contentBottom={
-        <RightRail />
       }
     />
   );
