+++ src/pages/research/collections/online-resources/featured/[slug].js
@@ -1,7 +1,6 @@
 import React, { Fragment, useEffect } from 'react';
 // Next
 import Router, { useRouter } from 'next/router';
-import Link from 'next/link';
 // Apollo
 import { getDataFromTree } from '@apollo/client/react/ssr';
 import { useQuery } from '@apollo/client';
@@ -16,12 +15,7 @@ import {
 import { withRedux } from './../../../../../redux/withRedux';
 // Components
 import { SkeletonLoader } from '@nypl/design-system-react-components';
-import PageContainer from './../../../../../components/shared/layouts/PageContainer';
-import RightRail from './../../../../../components/shared/RightRail';
-import SearchHeader from './../../../../../components/shared/SearchHeader';
-import SearchForm from './../../../../../components/online-resources/SearchForm';
-import SidebarMenus from './../../../../../components/online-resources/SidebarMenus';
-import Hero from './../../../../../components/online-resources/Hero';
+import PageContainer from './../../../../../components/online-resources/layouts/PageContainer';
 import SearchResults from './../../../../../components/online-resources/SearchResults';
 // Utils
 import { ONLINE_RESOURCES_BASE_PATH } from './../../../../../utils/config';
@@ -61,20 +55,8 @@ function FeaturedResourceTopicSlug() {
   if (loading || !data) {
     return (
       <PageContainer
-        wrapperClass='nypl--research'
-        contentHeader={
-          <Fragment>
-            <Hero />
-            <SearchHeader>
-              <SearchForm />
-            </SearchHeader>
-          </Fragment>
-        }
         contentPrimary={
           <SkeletonLoader />
-        }
-        contentBottom={
-          <RightRail />
         }
       />
     );
@@ -87,30 +69,11 @@ function FeaturedResourceTopicSlug() {
         description: `${data.resourceTopic.name}`,
         url: `https://www.nypl.org${ONLINE_RESOURCES_BASE_PATH}/${slug}`
       }}
-      wrapperClass='nypl--research'
-      contentHeader={
-        <Fragment>
-          <Hero />
-          <SearchHeader>
-            <SearchForm />
-          </SearchHeader>
-        </Fragment>
-      }
       contentPrimary={
-        <Fragment>
-          <SearchResults
-            resourceTopicId={data.resourceTopic.tid}
-            resourceTopicTitle={data.resourceTopic.name}
-          />
-        </Fragment>
-      }
-      showSidebar={true}
-      sidebarSide='right'
-      contentSecondary={
-        <SidebarMenus />
-      }
-      contentBottom={
-        <RightRail />
+        <SearchResults
+          resourceTopicId={data.resourceTopic.tid}
+          resourceTopicTitle={data.resourceTopic.name}
+        />
       }
     />
   );
