+++ src/pages/research/collections/online-resources/index.js
@@ -5,14 +5,9 @@ import { withApollo } from './../../../../apollo/client/withApollo';
 // Redux
 import { withRedux } from './../../../../redux/withRedux';
 // Components
-import PageContainer from './../../../../components/shared/layouts/PageContainer';
-import RightRail from './../../../../components/shared/RightRail';
-import SearchHeader from './../../../../components/shared/SearchHeader';
-import SearchForm from './../../../../components/online-resources/SearchForm';
+import PageContainer from './../../../../components/online-resources/layouts/PageContainer';
 import ResourceTopics from './../../../../components/online-resources/ResourceTopics';
 import MostPopularResources from './../../../../components/online-resources/MostPopularResources';
-import SidebarMenus from './../../../../components/online-resources/SidebarMenus';
-import Hero from './../../../../components/online-resources/Hero';
 // Utils
 import { ONLINE_RESOURCES_BASE_PATH } from './../../../../utils/config';
 
@@ -24,28 +19,11 @@ function OnlineResources() {
         description: 'Hello welcome to the NYPL!',
         url: `https://www.nypl.org${ONLINE_RESOURCES_BASE_PATH}`
       }}
-      wrapperClass='nypl--research'
-      contentHeader={
-        <Fragment>
-          <Hero />
-          <SearchHeader>
-            <SearchForm />
-          </SearchHeader>
-        </Fragment>
-      }
       contentPrimary={
         <Fragment>
           <ResourceTopics />
           <MostPopularResources />
         </Fragment>
-      }
-      showSidebar={true}
-      sidebarSide='right'
-      contentSecondary={
-        <SidebarMenus />
-      }
-      contentBottom={
-        <RightRail />
       }
     />
   );
