+++ src/pages/research/collections/online-resources/search.js
@@ -5,13 +5,8 @@ import { withApollo } from '../../../../apollo/client/withApollo';
 // Redux
 import { withRedux } from '../../../../redux/withRedux';
 // Components
-import PageContainer from '../../../../components/shared/layouts/PageContainer';
-import RightRail from '../../../../components/shared/RightRail';
-import SearchHeader from '../../../../components/shared/SearchHeader';
-import SearchForm from '../../../../components/online-resources/SearchForm';
+import PageContainer from './../../../../components/online-resources/layouts/PageContainer';
 import SearchResults from '../../../../components/online-resources/SearchResults';
-import SidebarMenus from '../../../../components/online-resources/SidebarMenus';
-import Hero from '../../../../components/online-resources/Hero';
 // Utils
 import { ONLINE_RESOURCES_BASE_PATH } from './../../../../utils/config';
 
@@ -19,31 +14,12 @@ function OnlineResourcesSearchPage() {
   return (
     <PageContainer
       metaTags={{
-        title: 'Online Resources',
+        title: 'Online Resources Search',
         description: 'Hello welcome to the NYPL!',
         url: `https://www.nypl.org${ONLINE_RESOURCES_BASE_PATH}/search`
       }}
-      wrapperClass='nypl--research'
-      contentHeader={
-        <Fragment>
-          <Hero />
-          <SearchHeader>
-            <SearchForm />
-          </SearchHeader>
-        </Fragment>
-      }
       contentPrimary={
-        <Fragment>
-          <SearchResults />
-        </Fragment>
-      }
-      showSidebar={true}
-      sidebarSide='right'
-      contentSecondary={
-        <SidebarMenus />
-      }
-      contentBottom={
-        <RightRail />
+        <SearchResults />
       }
     />
   );
